+++ posthog/tasks/exports/image_exporter.py
@@ -17,7 +17,7 @@ from statshog.defaults.django import statsd
 from webdriver_manager.chrome import ChromeDriverManager
 from webdriver_manager.core.utils import ChromeType
 
-from posthog.caching.update_caching_state import synchronously_update_cache
+from posthog.caching.insight_cache import synchronously_update_cache
 from posthog.logging.timing import timed
 from posthog.models.exported_asset import ExportedAsset, get_public_access_token, save_content
 from posthog.utils import absolute_uri
